requirements.txt
parsel
requests
tqdm
telebot
aiogram
python-telegram-bot
numpy
tqdm
parsel
